refactor(algorithms): refactor the code, now the project consists only in one maven module, the scopes are added to the dependencies
