+++ .github/workflows/unit-tests.yml
@@ -29,7 +29,7 @@ jobs:
           restore-keys: ${{ runner.os }}-maven
 
       - name: Install dependencies
-        run: mvn install -D skipTests
+        run: mvn install -D skipTests -D skipITs
 
       - name: Run unit tests
         run: mvn test
