--- types/bubble/src/test/java/xyz/zalaya/BubbleSortTest.java
@@ -1,13 +1,0 @@
-package xyz.zalaya;
-
-import xyz.zalaya.models.SortingAlgorithm;
-import xyz.zalaya.models.SortingAlgorithmTest;
-
-public class BubbleSortTest extends SortingAlgorithmTest {
-
-    @Override
-    public SortingAlgorithm getSortingAlgorithm() {
-        return new BubbleSort();
-    }
-
-}
