--- types/pom.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project
-    xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
->
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>xyz.zalaya</groupId>
-        <artifactId>sorting-algorithms</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>types</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>selection</module>
-        <module>bubble</module>
-        <module>insertion</module>
-    </modules>
-
-    <dependencies>
-        <dependency>
-            <groupId>xyz.zalaya</groupId>
-            <artifactId>common</artifactId>
-            <version>0.1.0-SNAPSHOT</version>
-        </dependency>
-    </dependencies>
-</project>
