--- types/selection/src/test/java/xyz/zalaya/SelectionSortTest.java
@@ -1,13 +1,0 @@
-package xyz.zalaya;
-
-import xyz.zalaya.models.SortingAlgorithm;
-import xyz.zalaya.models.SortingAlgorithmTest;
-
-public class SelectionSortTest extends SortingAlgorithmTest {
-
-    @Override
-    public SortingAlgorithm getSortingAlgorithm() {
-        return new SelectionSort();
-    }
-
-}
